+++ qt/python/mantidqt/MPLwidgets.py
@@ -5,7 +5,7 @@
 #     & Institut Laue - Langevin
 # SPDX - License - Identifier: GPL - 3.0 +
 from __future__ import (absolute_import, division, print_function)
-from gui_helper import set_matplotlib_backend
+from .gui_helper import set_matplotlib_backend
 
 backend = set_matplotlib_backend()  # must be at the top of this file
 if backend == 'Qt4Agg':
