+++ qt/python/mantidqt/gui_helper.py
