+++ qt/python/mantidqt/widgets/sliceviewer/view.py
@@ -10,8 +10,8 @@
 from __future__ import (absolute_import, division, print_function)
 from qtpy.QtWidgets import QWidget, QVBoxLayout
 from qtpy.QtCore import Qt
+from mantidqt.MPLwidgets import FigureCanvas, NavigationToolbar2QT as NavigationToolbar
 from matplotlib.figure import Figure
-from MPLwidgets import FigureCanvas, NavigationToolbar2QT as NavigationToolbar
 from .dimensionwidget import DimensionWidget
