+++ scripts/FilterEvents.py
@@ -7,7 +7,7 @@
 # pylint: disable=invalid-name
 from __future__ import (absolute_import, division, print_function)
 import sys
-from gui_helper import set_matplotlib_backend, get_qapplication
+from mantidqt.gui_helper import set_matplotlib_backend, get_qapplication
 set_matplotlib_backend()  # must be called before anything tries to use matplotlib
 from FilterEvents import eventFilterGUI  # noqa
