+++ scripts/QECoverage.py
@@ -7,12 +7,12 @@
 # pylint: disable=line-too-long, too-many-instance-attributes, invalid-name, missing-docstring, too-many-statements
 # pylint: disable= too-many-branches, no-self-use
 from __future__ import (absolute_import, division, print_function)
-from gui_helper import get_qapplication, show_interface_help
+from mantidqt.gui_helper import get_qapplication, show_interface_help
 import numpy as np
 import os
 import mantid
 from qtpy import QtWidgets
-from MPLwidgets import *
+from mantidqt.MPLwidgets import *
 import matplotlib
 from matplotlib.figure import Figure
 from scipy import constants
