+++ scripts/TofConverter/converterGUI.py
@@ -11,7 +11,7 @@ from qtpy.QtGui import QDoubleValidator
 from qtpy import QtCore
 import os
 from mantid.kernel import Logger
-from gui_helper import show_interface_help
+from mantidqt.gui_helper import show_interface_help
 import math
 import TofConverter.convertUnits
